+++ CHANGELOG.md
@@ -1137,6 +1137,7 @@ activity and on its child transactions (which can be actuals, refunds, and adjus
 
 - Add Level B ISPF ODA and non-ODA activity bulk uploads
 - Add Level C ISPF ODA activity bulk uploads
+- Hide the ISPF templates for level B bulk uploads when the ISPF feature flag is enabled
 
 [unreleased]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-122...HEAD
 [release-122]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-121...release-122